fix. step3 피드백 반영(random tictactoc 에서 사용하는 함수들 의존성 주입)
